+++ project/src/components/favorites-button/types.ts
@@ -1,0 +1,9 @@
+type FavoritesButtonProps = {
+  offerId: number;
+  isFavorite: boolean;
+  width: string;
+  height: string;
+  className: string;
+}
+
+export type {FavoritesButtonProps};
